chore: Moved tools and some cross compiling logic into testsuite
