+++ tests/testsuite/build.rs
@@ -5,6 +5,7 @@ use std::fs;
 use std::io::Read;
 use std::process::Stdio;
 
+use crate::utils::tools;
 use cargo::core::compiler::UserIntent;
 use cargo::core::Shell;
 use cargo::core::Workspace;
@@ -18,7 +19,7 @@ use cargo_test_support::str;
 use cargo_test_support::{
     basic_bin_manifest, basic_lib_manifest, basic_manifest, cargo_exe, cargo_process, git,
     is_nightly, main_file, paths, process, project, rustc_host, sleep_ms, symlink_supported, t,
-    tools, Execs, ProjectBuilder,
+    Execs, ProjectBuilder,
 };
 use cargo_util::paths::dylib_path_envvar;
