+++ tests/testsuite/cache_messages.rs
@@ -1,8 +1,8 @@
 //! Tests for caching compiler diagnostics.
 
+use crate::utils::tools;
 use cargo_test_support::prelude::*;
 use cargo_test_support::str;
-use cargo_test_support::tools;
 use cargo_test_support::{basic_manifest, is_coarse_mtime, project, registry::Package, sleep_ms};
 
 use super::messages::raw_rustc_output;
