+++ tests/testsuite/cargo_alias_config.rs
@@ -2,9 +2,9 @@
 
 use std::env;
 
+use crate::utils::tools::echo_subcommand;
 use cargo_test_support::prelude::*;
 use cargo_test_support::str;
-use cargo_test_support::tools::echo_subcommand;
 use cargo_test_support::{basic_bin_manifest, project};
 
 #[cargo_test]
