+++ tests/testsuite/features2.rs
@@ -2,7 +2,8 @@
 
 use std::fs::File;
 
-use cargo_test_support::cross_compile::{self, alternate};
+use crate::utils::cross_compile::disabled as cross_compile_disabled;
+use cargo_test_support::cross_compile::alternate;
 use cargo_test_support::paths;
 use cargo_test_support::prelude::*;
 use cargo_test_support::publish::validate_crate_contents;
@@ -263,7 +264,7 @@ common
 #[cargo_test]
 fn itarget_proc_macro() {
     // itarget inside a proc-macro while cross-compiling
-    if cross_compile::disabled() {
+    if cross_compile_disabled() {
         return;
     }
     Package::new("hostdep", "1.0.0").publish();
@@ -1312,7 +1313,7 @@ fn has_dev_dep_for_test() {
 #[cargo_test]
 fn build_dep_activated() {
     // Build dependencies always match the host for [target.*.build-dependencies].
-    if cross_compile::disabled() {
+    if cross_compile_disabled() {
         return;
     }
     Package::new("somedep", "1.0.0")
