+++ tests/testsuite/fetch.rs
@@ -1,5 +1,6 @@
 //! Tests for the `cargo fetch` command.
 
+use crate::utils::cross_compile::disabled as cross_compile_disabled;
 use cargo_test_support::prelude::*;
 use cargo_test_support::registry::Package;
 use cargo_test_support::rustc_host;
@@ -17,7 +18,7 @@ fn no_deps() {
 
 #[cargo_test]
 fn fetch_all_platform_dependencies_when_no_target_is_given() {
-    if cross_compile::disabled() {
+    if cross_compile_disabled() {
         return;
     }
 
@@ -69,7 +70,7 @@ fn fetch_all_platform_dependencies_when_no_target_is_given() {
 
 #[cargo_test]
 fn fetch_platform_specific_dependencies() {
-    if cross_compile::disabled() {
+    if cross_compile_disabled() {
         return;
     }
