+++ tests/testsuite/fix.rs
@@ -1,5 +1,6 @@
 //! Tests for the `cargo fix` command.
 
+use crate::utils::tools;
 use cargo::core::Edition;
 use cargo_test_support::compare::assert_e2e;
 use cargo_test_support::git::{self, init};
@@ -7,7 +8,6 @@ use cargo_test_support::paths;
 use cargo_test_support::prelude::*;
 use cargo_test_support::registry::{Dependency, Package};
 use cargo_test_support::str;
-use cargo_test_support::tools;
 use cargo_test_support::{basic_manifest, is_nightly, project};
 
 #[cargo_test]
