+++ tests/testsuite/fix_n_times.rs
@@ -17,8 +17,9 @@
 use std::path::PathBuf;
 use std::sync::{Mutex, OnceLock};
 
+use crate::utils::tools;
 use cargo_test_support::prelude::*;
-use cargo_test_support::{basic_manifest, paths, project, str, tools, Execs};
+use cargo_test_support::{basic_manifest, paths, project, str, Execs};
 
 /// The action that the `rustc` shim should take in the current sequence of
 /// events.
