+++ tests/testsuite/rustdoc.rs
@@ -4,6 +4,8 @@ use cargo_test_support::prelude::*;
 use cargo_test_support::str;
 use cargo_test_support::{basic_manifest, cross_compile, project};
 
+use crate::utils::cross_compile::disabled as cross_compile_disabled;
+
 #[cargo_test]
 fn rustdoc_simple() {
     let p = project().file("src/lib.rs", "").build();
@@ -280,7 +282,7 @@ fn proc_macro_crate_type() {
 
 #[cargo_test]
 fn rustdoc_target() {
-    if cross_compile::disabled() {
+    if cross_compile_disabled() {
         return;
     }
