+++ tests/testsuite/utils/tools.rs
@@ -1,6 +1,6 @@
 //! Common executables that can be reused by various tests.
 
-use crate::{basic_manifest, paths, project, Project};
+use cargo_test_support::{basic_manifest, paths, project, Project};
 use std::path::{Path, PathBuf};
 use std::sync::Mutex;
 use std::sync::OnceLock;
